# docs/yaml/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Demonstration API
  description: A simple blog API.
  contact:
    name: Dennis Au-Yeung
    email: 217013622@stu.vtc.edu.hk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://BScCOM-Blog-API-example.wongazta.repl.co'
    description: Development server
...
paths:
  /auth:
    post:
      tags:
        - Authentication
      description: API endpoints for user login by using basic authentication
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:      # Response body contents
                  type: object
                examples:
                  expect:  
                    externalValue: '../responseJson/login.json'
        '401':
          description: Not authenticated   
        '500':
          description: Server error   
  /google/token:
    post:
      tags:
        - Authentication
      description: API endpoints for user login by using google token
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:      # Response body contents
                  type: object
                examples:
                  expect:  
                    externalValue: '../responseJson/login.json'          
        '401':
          description: Not authenticated 
        '500':
          description: Server error     
...